backend/requirements.txt
flask
python-dotenv
supabase
pillow
torch
boto3
ultralytics
